+++ cmd/rp-indexer/main.go
@@ -3,86 +3,81 @@ package main
 import (
 	"database/sql"
 	"os"
-	"time"
+	"os/signal"
+	"syscall"
 
 	"github.com/evalphobia/logrus_sentry"
 	_ "github.com/lib/pq"
 	"github.com/nyaruka/ezconf"
+	indexer "github.com/nyaruka/rp-indexer"
 	"github.com/nyaruka/rp-indexer/indexers"
 	log "github.com/sirupsen/logrus"
 )
 
-type config struct {
-	ElasticURL string `help:"the url for our elastic search instance"`
-	DB         string `help:"the connection string for our database"`
-	Index      string `help:"the alias for our contact index"`
-	Poll       int    `help:"the number of seconds to wait between checking for updated contacts"`
-	Rebuild    bool   `help:"whether to rebuild the index, swapping it when complete, then exiting (default false)"`
-	Cleanup    bool   `help:"whether to remove old indexes after a rebuild"`
-	LogLevel   string `help:"the log level, one of error, warn, info, debug"`
-	SentryDSN  string `help:"the sentry configuration to log errors to, if any"`
-}
-
 func main() {
-	config := config{
-		ElasticURL: "http://localhost:9200",
-		DB:         "postgres://localhost/temba?sslmode=disable",
-		Index:      "contacts",
-		Poll:       5,
-		Rebuild:    false,
-		Cleanup:    false,
-		LogLevel:   "info",
-	}
-	loader := ezconf.NewLoader(&config, "indexer", "Indexes RapidPro contacts to ElasticSearch", []string{"indexer.toml"})
+	cfg := indexer.NewDefaultConfig()
+	loader := ezconf.NewLoader(cfg, "indexer", "Indexes RapidPro contacts to ElasticSearch", []string{"indexer.toml"})
 	loader.MustLoad()
 
 	// configure our logger
 	log.SetOutput(os.Stdout)
 	log.SetFormatter(&log.TextFormatter{})
 
-	level, err := log.ParseLevel(config.LogLevel)
+	level, err := log.ParseLevel(cfg.LogLevel)
 	if err != nil {
 		log.Fatalf("Invalid log level '%s'", level)
 	}
 	log.SetLevel(level)
 
 	// if we have a DSN entry, try to initialize it
-	if config.SentryDSN != "" {
-		hook, err := logrus_sentry.NewSentryHook(config.SentryDSN, []log.Level{log.PanicLevel, log.FatalLevel, log.ErrorLevel})
+	if cfg.SentryDSN != "" {
+		hook, err := logrus_sentry.NewSentryHook(cfg.SentryDSN, []log.Level{log.PanicLevel, log.FatalLevel, log.ErrorLevel})
 		hook.Timeout = 0
 		hook.StacktraceConfiguration.Enable = true
 		hook.StacktraceConfiguration.Skip = 4
 		hook.StacktraceConfiguration.Context = 5
 		if err != nil {
-			log.Fatalf("invalid sentry DSN: '%s': %s", config.SentryDSN, err)
+			log.Fatalf("invalid sentry DSN: '%s': %s", cfg.SentryDSN, err)
 		}
 		log.StandardLogger().Hooks.Add(hook)
 	}
 
-	db, err := sql.Open("postgres", config.DB)
+	db, err := sql.Open("postgres", cfg.DB)
 	if err != nil {
-		log.Fatal(err)
+		log.Fatalf("unable to connect to database")
 	}
 
-	ci := indexers.NewContactIndexer(config.ElasticURL, config.Index, 500)
+	idxrs := []indexers.Indexer{
+		indexers.NewContactIndexer(cfg.ElasticURL, cfg.Index, 500),
+	}
+
+	if cfg.Rebuild {
+		// if rebuilding, just do a complete index and quit. In future when we support multiple indexers,
+		// the rebuild argument can be become the name of the index to rebuild, e.g. --rebuild=contacts
+		idxr := idxrs[0]
+		if _, err := idxr.Index(db, true, cfg.Cleanup); err != nil {
+			log.WithField("indexer", idxr.Name()).WithError(err).Fatal("error during rebuilding")
+		}
+	} else {
+		d := indexer.NewDaemon(cfg, db, idxrs)
+		d.Start()
+
+		handleSignals(d)
+	}
+}
+
+// handleSignals takes care of trapping quit, interrupt or terminate signals and doing the right thing
+func handleSignals(d *indexer.Daemon) {
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
 
 	for {
-		_, err := ci.Index(db, config.Rebuild, config.Cleanup)
-
-		if err != nil {
-			if config.Rebuild {
-				log.WithField("index", config.Index).WithError(err).Fatal("error during rebuilding")
-			} else {
-				log.WithField("index", config.Index).WithError(err).Error("error during indexing")
-			}
+		sig := <-sigs
+		switch sig {
+		case syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT:
+			log.WithField("signal", sig).Info("received exit signal, exiting")
+			d.Stop()
+			return
 		}
-
-		// if we were rebuilding then we're done
-		if config.Rebuild {
-			os.Exit(0)
-		}
-
-		// sleep a bit before starting again
-		time.Sleep(time.Second * 5)
 	}
 }